+++ angular/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 6.0.1-dev
+
+* Update `package:analyzer` dependency to `^0.40.0`.
+
 ## 6.0.0
 
 ### New features