+++ angular_compiler/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1-dev
 
 * Removed invalid link from the OnPush compatibility warning.
+* Update `package:analyzer` dependency to `^0.40.0`.
 
 ## 1.0.0